PERA-1357 Rename indexer api service and move it to foundation
